+++ tests/testthat/test-cnd-message.R
@@ -333,11 +333,50 @@ test_that("special syntax calls handle edge cases", {
   expect_equal(error_call_as_string(quote(base::`+`(1, 2))), "+")
 })
 
-test_that("can print message with prefix", {
-  foo <- error_cnd("foo", message = "Foo")
-  bar <- error_cnd("bar", message = "Bar", parent = foo)
-  expect_snapshot({
+test_that("can print message with and without prefix", {
+  expect_snapshot(cran = TRUE, {
+    foo <- error_cnd(
+      "foo",
+      message = "Parent message.",
+      body = c("*" = "Bullet 1.", "*" = "Bullet 2."),
+      use_cli_format = TRUE
+    )
+    bar <- error_cnd(
+      "bar",
+      message = "Message.",
+      body = c("*" = "Bullet A.", "*" = "Bullet B."),
+      parent = foo,
+      use_cli_format = TRUE
+    )
+
     writeLines(cnd_message(foo, prefix = TRUE))
     writeLines(cnd_message(bar, prefix = TRUE))
+
+    writeLines(cnd_message(foo, prefix = FALSE))
+    writeLines(cnd_message(bar, prefix = FALSE))
+  })
+})
+
+test_that("can print message without inheritance", {
+  expect_snapshot(cran = TRUE, {
+    foo <- error_cnd(
+      "foo",
+      message = "Parent message.",
+      body = c("*" = "Bullet 1.", "*" = "Bullet 2."),
+      use_cli_format = TRUE
+    )
+    bar <- error_cnd(
+      "bar",
+      message = "Message.",
+      body = c("*" = "Bullet A.", "*" = "Bullet B."),
+      parent = foo,
+      use_cli_format = TRUE
+    )
+
+    writeLines(cnd_message(foo, inherit = FALSE, prefix = TRUE))
+    writeLines(cnd_message(bar, inherit = FALSE, prefix = TRUE))
+
+    writeLines(cnd_message(foo, inherit = FALSE, prefix = FALSE))
+    writeLines(cnd_message(bar, inherit = FALSE, prefix = FALSE))
   })
 })